uqcsbot: tidied up status reacts functions
